Update cpu.cycles in interrupt_guard()
The STI instruction modifies cpu.cycles_to_run, which is used to keep track of the number of cycles that have been executed. Updating this value changes the value returned by cpu_get_cycles().
Updating cpu.cycles mitigates the change that modifying cpu.cycles_to_run could have caused so that time seems to be constantly increasing. It's a hack, but it works.
I believe this confused the PIT, and it began firing interrupts much too slowly.

I accidentally deleted this line of code when tidying up the source. Fixing this bug allows Windows 95 to boot again.