Fix PRDT remapping for PCI IDE
The PRDT can be remapped to different ports by changing BAR4 of the PCI configuration space. The problem was that BAR4 was a 32-bit quantity and the PCI bus handled configuration space modifications by bytes only. This meant that the PDRT was being remapped four times, often with the wrong address.
Now the PRDT is only modified when address == 0x23, the last byte of BAR4.