Fix ESP handling for pop [mem]
pop [mem] first pops the value off the stack, computes the effective address (important for instructions like pop [esp]) and writes it to memory.
In rare cases, the pop doesn't fault, but the write does. In that case, the ESP being pushed to stack was the one *following* the pop, not before it, as it should be.

With this fix, KolibriOS boots. It reverses the previous commit.