Fix method signatures
Windows 98 would crash with a #DE exception because Emscripten would mistakenly sign extend the second argument of cpu_muldiv16 because the method signature in opcodes.h was "int cpu_muldiv(int op, uint16_t src)"
It would read from the HEAP16 (signed 16-bit integers) heap.

The actual function was int cpu_muldiv16(int op, uint32_t src), so the sign extension would not be truncated. This caused MUL to act like IMUL, which caused Windows 98 to crash with a divide-by-zero fault later on.
With this bugfix, Windows 98 boots in the Emscripten version as well as the native version.