Fix outport and fix IRQ routing. OS/2 Warp 3 boots now
The 8042 "OutPort" command (0xD2) was sending the LSB of the command byte (0xD2 in this case, so 0x00 was sent) instead of the command byte itself.
Apparently, the PS/2 controller can send IRQs even when the clock is turned off. To be quite honest, I'm not sure what the clock does, but I believe it has something to do with how fast the queue is refilled. This is a non-issue in Halfix since our policy is to ignore what the OS wants and refill the queue as fast as possible.