Complete rewrite of configuration parser.
The old version was clunky and was buggy in slight ways (i.e. the "driver" directive always had to appear in front of the "file" one).
Crucially, it did not allow you to have drives with no media attached to them (which is required for Windows NT 4.0 to boot).

This version rectifies these mistakes, at least on the native version. Work on the Emscripten version is pending.