Fix split-screen bug
The VGA emulation renders the screen scanline-by-scanline. It's CPU-intensive but a fairly accurate representation of what a VGA controller actually does. Because it's so CPU intensive, we only update half of the screen at once. We call vga_update() 60 times a second to achieve a frame-rate of 30 FPS.
A further optimization is to not re-render the scanlines if VGA memory has not been touched since the last time vga_update() was called (VESA graphics modes go further and determine which scanlines should be updated). When we finished rendering the last scanline, we'd tell the VGA emulator that what's on the screen is in line with what's in video memory, so no further rendering is necessary for now.
The annoying "split screen" effect occurred when video RAM was modified in between the half-screen renders. Here's a short description of what would happen:

 - The top half of the screen would be redrawn
 - The OS would change the top half of the screen
 - The bottom half of the screen would be redrawn
 - The memory-modified flag would be cleared, leaving the top half containing outdated information.

It was especially apparent in real mode during text mode scrolling.

This commit adds a separate flag for both the top and bottom halves of the screen.