Fix cross-sector ATAPI reads (#2)
Cross-sector ATAPI reads were problematic because we determined how many bytes of the previous frame to save based on the transfer count. This led to cases where the first few bytes of the sector would be read-read under certain circumstances.
Now we determine how many bytes to save based on the total bytes transferred. This fixes TinyCore Linux and allows it to boot to desktop.